TestFixAndCrit

+++ UnityProject/Assets/Shaders/CritStateShader.shader
@@ -17,17 +17,17 @@
 
 				// frag shaders data
 				uniform sampler2D _MainTex;
-				uniform int _CurrentHealth = 0;
+				uniform float _CurrentHealth = 0.0f;
 					
 				static const float _GrainSpeed = 0.1f;
-				static const int vignetteStart = 25;
-				static const int vignetteEnd = -100;
-				static const int greyScaleStart = 25;
-				static const int greScaleEnd = -15;
-				static const int redTintStart = 10;
-				static const int redTintEnd = -60;
-				static const int darkeningStart = 0;
-				static const int darkeningEnd = -100;
+				static const int vignetteStart = 25.0f;
+				static const int vignetteEnd = -100.0f;
+				static const int greyScaleStart = 25.0f;
+				static const int greScaleEnd = -15.0f;
+				static const int redTintStart = 10.0f;
+				static const int redTintEnd = -60.0f;
+				static const int darkeningStart = 0.0f;
+				static const int darkeningEnd = -100.0f;
 					
 				static const float3 BT709 = float3(0.2126f, 0.7152f, 0.0722f);
 				static const float noiseIntensity = 0.04f;
